refactor: Move RaptorConstants into raptor.api.model package
