+++ src/main/java/org/opentripplanner/raptor/path/PathBuilderLeg.java
@@ -3,8 +3,8 @@ package org.opentripplanner.raptor.path;
 import java.util.function.Predicate;
 import javax.annotation.Nullable;
 import org.opentripplanner.framework.time.TimeUtils;
-import org.opentripplanner.raptor.api.RaptorConstants;
 import org.opentripplanner.raptor.api.model.RaptorAccessEgress;
+import org.opentripplanner.raptor.api.model.RaptorConstants;
 import org.opentripplanner.raptor.api.model.RaptorConstrainedTransfer;
 import org.opentripplanner.raptor.api.model.RaptorTransfer;
 import org.opentripplanner.raptor.api.model.RaptorTransferConstraint;
