+++ src/main/java/org/opentripplanner/raptor/rangeraptor/debug/DebugHandlerPathAdapter.java
@@ -1,7 +1,7 @@
 package org.opentripplanner.raptor.rangeraptor.debug;
 
 import java.util.List;
-import org.opentripplanner.raptor.api.RaptorConstants;
+import org.opentripplanner.raptor.api.model.RaptorConstants;
 import org.opentripplanner.raptor.api.path.EgressPathLeg;
 import org.opentripplanner.raptor.api.path.RaptorPath;
 import org.opentripplanner.raptor.api.request.DebugRequest;
