+++ src/main/java/org/opentripplanner/raptor/rangeraptor/multicriteria/arrivals/c2/AccessStopArrivalC2.java
@@ -2,9 +2,9 @@ package org.opentripplanner.raptor.rangeraptor.multicriteria.arrivals.c2;
 
 import static org.opentripplanner.raptor.api.model.PathLegType.ACCESS;
 
-import org.opentripplanner.raptor.api.RaptorConstants;
 import org.opentripplanner.raptor.api.model.PathLegType;
 import org.opentripplanner.raptor.api.model.RaptorAccessEgress;
+import org.opentripplanner.raptor.api.model.RaptorConstants;
 import org.opentripplanner.raptor.api.model.RaptorTripSchedule;
 import org.opentripplanner.raptor.api.view.AccessPathView;
 import org.opentripplanner.raptor.spi.RaptorCostCalculator;
