+++ src/main/java/org/opentripplanner/raptor/spi/RaptorBoardOrAlightEvent.java
@@ -2,7 +2,7 @@ package org.opentripplanner.raptor.spi;
 
 import java.util.function.Consumer;
 import javax.annotation.Nonnull;
-import org.opentripplanner.raptor.api.RaptorConstants;
+import org.opentripplanner.raptor.api.model.RaptorConstants;
 import org.opentripplanner.raptor.api.model.RaptorTransferConstraint;
 import org.opentripplanner.raptor.api.model.RaptorTripSchedule;
