+++ src/test/java/org/opentripplanner/raptor/RaptorArchitectureTest.java
@@ -58,7 +58,6 @@ public class RaptorArchitectureTest {
   );
 
   @Test
-  @Disabled
   void enforcePackageDependenciesRaptorAPI() {
     API_MODEL.dependsOn(FRAMEWORK_UTILS).verify();
     API_PATH.dependsOn(FRAMEWORK_UTILS, API_MODEL).verify();
