+++ src/test/java/org/opentripplanner/raptor/_data/transit/TestAccessEgress.java
@@ -3,15 +3,15 @@ package org.opentripplanner.raptor._data.transit;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNull;
 import static org.junit.jupiter.api.Assertions.assertTrue;
-import static org.opentripplanner.raptor.api.RaptorConstants.SECONDS_IN_A_DAY;
+import static org.opentripplanner.raptor.api.model.RaptorConstants.SECONDS_IN_A_DAY;
 import static org.opentripplanner.routing.algorithm.raptoradapter.transit.cost.RaptorCostConverter.toRaptorCost;
 
 import java.util.ArrayList;
 import java.util.Collection;
 import java.util.List;
 import org.opentripplanner.framework.time.TimeUtils;
-import org.opentripplanner.raptor.api.RaptorConstants;
 import org.opentripplanner.raptor.api.model.RaptorAccessEgress;
+import org.opentripplanner.raptor.api.model.RaptorConstants;
 
 /**
  * Simple implementation for {@link RaptorAccessEgress} for use in unit-tests.
